cleanup update-cdm-deliverable script calls inside fetch-dmp

+++ import-scripts/update-cdm-deliverable.sh
@@ -1,22 +1,55 @@
 #!/bin/bash
-source $PORTAL_HOME/scripts/filter-clinical-arg-functions.sh
+
+DELIVERED_SAMPLE_ATTRIBUTES="SAMPLE_ID PATIENT_ID CANCER_TYPE CANCER_TYPE_DETAILED"
 
 CLINICAL_SAMPLE_FILE=$1
 SEQ_DATE_FILE=$2
+STUDY_ID=$3
 
-DELIVERED_SAMPLE_ATTRIBUTES="SAMPLE_ID PATIENT_ID CANCER_TYPE CANCER_TYPE_DETAILED"
+echo "`date`: Starting CDM deliverable generation and upload"
+if [ ! -f $CLINICAL_SAMPLE_FILE ] || [ ! -f $SEQ_DATE_FILE ] ; then
+  echo "`date`: One or more supplied files ($CLINICAL_SAMPLE_FILE and $SEQ_DATE_FILE) do not exist, exiting..."
+  exit 1
+fi
+
+if [ ! -f $PORTAL_HOME/scripts/automation_env.sh ] || [ ! -f $PORTAL_HOME/scripts/filter-clinical-arg-functions.sh ] ; then
+  echo "`date`: Unable to locate automation_env and additional modules, exiting..."
+  exit 1
+fi
+
+source $PORTAL_HOME/scripts/automation_env.sh
+source $PORTAL_HOME/scripts/filter-clinical-arg-functions.sh
 
 TMP_SAMPLE_FILE=$(mktemp -q)
 TMP_PROCESSING_FILE=$(mktemp -q)
 CDM_DELIVERABLE=$(mktemp -q)
 
+# Copy sample file to tmp file since script overwrites existing file (don't want to overwrite DMP pipeline files)
+# Removes all clinical attributes except those specified in $DELIVERED_SAMPLE_ATTRIBUTES set at top
+# TMP_PROCESSING_FILE automatically removed // TODO: move TMP_FILE creation to filter_clinical function
 cp -a $CLINICAL_SAMPLE_FILE $TMP_SAMPLE_FILE
 filter_clinical_attribute_columns "$TMP_SAMPLE_FILE" "$DELIVERED_SAMPLE_ATTRIBUTES" "$TMP_PROCESSING_FILE"
+if [ $? -ne 0 ] ; then
+  echo "`date`: Failed to subset clinical sample file, exiting..."
+  exit 1
+fi
 
+# Combines filtered clinical sample file with seq data file and outputs to tmp file for upload
+# Uses left join -- SAMPLE_ID in the clinical_sample_file (first arg) will be valid keys
 $PYTHON3_BINARY $PORTAL_HOME/scripts/combine_files_py3.py -i "$TMP_SAMPLE_FILE" "$SEQ_DATE_FILE" -o "$CDM_DELIVERABLE" -c SAMPLE_ID -m left 
+if [ $? -ne 0 ] ; then
+  echo "`date`: Failed to combine files, exiting..."
+  exit 1
+fi
 
-aws s3 cp $CDM_DELIVERABLE s3://cdm-deliverable/averystest.txt --profile saml 
-cp $CDM_DELIVERABLE . 
+# Authenticate and upload into S3 bucket
+$PORTAL_HOME/scripts/authenticate_service_account.sh eks
+aws s3 cp $CDM_DELIVERABLE s3://cdm-deliverable/data_clinical_$STUDY_ID.txt --profile saml
+if [ $? -ne 0 ] ; then
+  echo "`date`: Failed to upload CDM deliverable to S3, exiting..."
+fi
 
 rm $TMP_SAMPLE_FILE
 rm $CDM_DELIVERABLE
+
+echo "`date`: CDM deliverable generation and upload complete"
